+++ packages/front-end/src/views/challenge/ChallengeView.tsx
@@ -5,9 +5,9 @@ import { useDidMount } from "rooks";
 import { Challenge, HttpChallengeService } from "../../api/challenge";
 import { HttpClient } from "../../api/http";
 import { HeadingGroup } from "../../components/heading-group";
-import { ChallengeResults } from "../../components/challenge/challenge-results";
+import { ChallengeResults } from "./components/challenge-results";
 import { ChallengeResult } from "../../api/challenge/ChallengeResult";
-import { ChallengeResultAction } from "../../components/challenge/challenge-results/ChallengeResultAction";
+import { ChallengeResultAction } from "./components/challenge-results/ChallengeResultAction";
 import { SessionPayload } from "../../api/session";
 import { HttpChallengeParticipantService } from "../../api/challenge/HttpChallengeParticipantService";
 import { ChallengeParticipantUpdateRequest } from "../../api/challenge/ChallengeParticipantUpdateRequest";
