+++ packages/front-end/src/views/challenge/components/challenge-results/ChallengeResults.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/variables.scss";
+@import "../../../../styles/variables.scss";
 
 .ChallengeResults {
   display: flex;
