+++ packages/front-end/src/views/challenge/components/challenge-results/EmptyChallengeResults.tsx
